backend/requirements.txt
Flask
python-dotenv
supabase
pandas
pypdf
beautifulsoup4
python-docx
PyYAML
python-pptx
langchain
langchain_groq
langchain_community
redis
faiss-cpu
scikit-learn
sentence-transformers
langchain-huggingface
tiktoken
zipfile